Moved some calculation tools to utility. Moved DropSizeDistribution to io folder. Modified ParsivelReader to remove a couple of bugs. Modified test to account for new path. Added a pydsd folder and __init__.py to map to pydisdrometer. Modified the setup.py to allow calling package by pydisdromter or pydsd.
