+++ pydisdrometer/tests/test_DSR.py
@@ -1,4 +1,4 @@
 import pydisdrometer
 
 def test_tb():
-    assert pydisdrometer.DSR.tb(2) > 0
+    assert pydisdrometer.utility.DSR.tb(2) > 0
