+++ setup.py
@@ -6,11 +6,11 @@ setup(
     author='Joseph C. Hardin',
     author_email='josephhardinee@gmail.com',
     packages=['pydisdrometer',
-              'pydisdrometer.aux_readers',
               'pydisdrometer.io',
               'pydisdrometer.partition',
               'pydisdrometer.plot',
-              'pydisdrometer.utility'],
+              'pydisdrometer.utility',
+              'pydsd',],
     url='http://pypi.python.org/pypi/PyDisdrometer/',
     license='LICENSE.txt',
     description='Python Disdrometer Processing',
